made another field unsigned

+++ src/EZButton.cpp
@@ -69,7 +69,7 @@ void EZButton::CheckButtons()
             }
             else if (millis() - _buttonDownMillis[i] > HoldTreshHold)
             {
-                int interval = (millis() - _buttonDownMillis[i]) / HoldInterval;
+                unsigned long interval = (millis() - _buttonDownMillis[i]) / HoldInterval;
                 if (interval > _lastHoldInterval[i])
                 {
                     _lastHoldInterval[i] = interval;
